+++ src/main/java/rentalstore/price/Price.java
@@ -1,11 +1,9 @@
 package rentalstore.price;
 
-import rentalstore.Movie;
-
 public abstract class Price {
     public abstract int getPriceCode();
     public abstract double getAmount(int dayRented);
-    public int getFrequentRenterPoints(int dayRented) {
-        return 1;
+    public double getFrequentRenterPoints(int dayRented) {
+        return 1.0;
     }
 }
